Make collsion bullet with enemy and obstacle and refactoring gun.h gun.c
