+++ HiddenPickle.vcxproj
@@ -202,6 +202,7 @@ xcopy "$(ProjectDir)Assets\*" "$(OutDir)Assets\" /s /r /y /q</Command>
     <ClCompile Include="exitPlace.c" />
     <ClCompile Include="footprint.c" />
     <ClCompile Include="gameManager.c" />
+    <ClCompile Include="gun.c" />
     <ClCompile Include="itemBox.c" />
     <ClCompile Include="main.c" />
     <ClCompile Include="obstacle.c" />
@@ -216,6 +217,7 @@ xcopy "$(ProjectDir)Assets\*" "$(OutDir)Assets\" /s /r /y /q</Command>
     <ClInclude Include="exitPlace.h" />
     <ClInclude Include="footprint.h" />
     <ClInclude Include="gameManager.h" />
+    <ClInclude Include="gun.h" />
     <ClInclude Include="itemBox.h" />
     <ClInclude Include="obstacle.h" />
     <ClInclude Include="player.h" />
