+++ itemBox.h
@@ -19,24 +19,13 @@
 
 #define MAX_BULLET 6
 #define MAX_BATTERY 100
-#define TIME_BULLLET 2
+
 
 typedef enum item {
 	KEY_Item,
 	BULLET_Item, // Resource of Gun
 	BATTERY_Item // Resource of Lamp
 } EItemType;
-
-typedef struct gun {
-	CP_Vector position_Bullet[MAX_BULLET];
-	CP_Vector vector_Shooting[MAX_BULLET];
-	float time_Shooting[MAX_BULLET];
-	int count_Bullet;
-	float radius_Bullet;
-	float speed_Bullet;
-	int attackPoint;
-	//float range; // Shooting Range
-} GUN;
 
 
 typedef struct item_box {
@@ -57,18 +46,5 @@ int get_Item_Type(ITEM_BOX* item_box);
 
 int isEmptyBox(ITEM_BOX* item_box);
 
-void init_Gun(GUN* gun);
-
-int getIndexEmptyBullet(GUN* gun);
-
-void addBullet_Gun(GUN* gun, CP_Vector startPosition, CP_Vector vectorShooting, int indexEmpty);
-
-void update_Gun(GUN* gun, float dt);
-
-int checkCollision_Bullet_Object(GUN* gun, CP_Vector positionObject, float radius);
-
 //void endBullet(GUN* gun);
 
-int isBulletShooting(float time_Shooting); 
-
-void printBullet(GUN* gun);
